Add Delete icon to favourite item list Change add fav button to Details page Changed fav icon
