+++ app/src/main/res/values-ja-ldltr/strings.xml
@@ -16,4 +16,5 @@
     <string name="select_app_language">アプリの言語</string>
     <string name="exit_confirmation_message">退出しますか?</string>
     <string name="progress_dialog_message">待っている......</string>
+    <string name="add_fav_confirmation_message">お気に入りに追加しますか?</string>
 </resources>
