+++ app/src/main/res/values/dimens.xml
@@ -29,4 +29,5 @@
 
     /* Action Bar Dimensions */
     <dimen name="action_bar_image_button_icon_padding">16dp</dimen>
+    <dimen name="image_button_size">30dp</dimen>
 </resources>
