+++ app/src/main/res/values/strings.xml
@@ -17,5 +17,6 @@
     <!-- Dialog messages -->
     <string name="exit_confirmation_message">Do you want to exit?</string>
     <string name="progress_dialog_message">Processing......</string>
+    <string name="add_fav_confirmation_message">Do you want to add to favourites?</string>
 
 </resources>
